add to Education Api update, delete and getById services

+++ src/main/java/com/umutdoruk/hrms/repository/EducationRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface EducationRepository extends JpaRepository<Education,Integer> {
+public interface EducationRepository extends JpaRepository<Education,Long> {
     List<Education> findAllByOrderByGraduationDateAsc();
     List<Education> findAllByOrderByGraduationDateDesc();
 }
